requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.18.0
Pillow>=7.0.0
tqdm>=4.40.0
scikit-image>=0.16.0
scipy>=1.4.0
